requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
pytest
httpx
coverage
alembic
python-dotenv
pytest-cov
locust